+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # DEV WIP CHANGES
+<<<<<<< HEAD
+- Added functionality to automatically resize the character and item sheets
+=======
 - added Debug Elements setting for developers
 - removed hp, humanity, and emp calculations for mooks
+>>>>>>> origin/dev
 
 # DEV WIP BUG FIXES
 - fixed #189